requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
jinja2==3.1.6
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# PDF Processing
PyMuPDF==1.23.8
pdfplumber==0.10.3

# Vector Database
pinecone-client==3.0.3

# Embeddings & AI
sentence-transformers==2.7.0
torch==2.1.2
transformers==4.40.0
google-generativeai==0.8.3
tiktoken==0.5.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic[email]==2.5.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
numpy==1.24.4
pandas==2.1.4
tenacity==8.2.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1